requirements.txt
requests==2.13.0
python-dateutil==2.6.0